requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# OpenAI API
openai==1.55.3

# Data Validation
pydantic==2.10.2

# Fine-tuning (Hugging Face)
datasets==3.1.0

# Additional dependencies
python-multipart==0.0.19
python-dotenv==1.0.1
